closes https://github.com/tcurdt/jdeb/issues/214 declared some dependencies as provided upgraded dependencies added prerequisites to pom

+++ src/main/java/org/vafer/jdeb/DataConsumer.java
@@ -25,7 +25,7 @@ import java.io.InputStream;
  */
 public interface DataConsumer {
 
-    void onEachDir( String dirname, String linkname, String user, int uid, String group, int gid, int mode, long size ) throws IOException;
+    void onEachDir( String dirname, String linkname, String user, long uid, String group, long gid, int mode, long size ) throws IOException;
 
     void onEachFile( InputStream input, TarArchiveEntry entry ) throws IOException;
